+++ src/TestTrackingDiagrams.Extensions.Kafka/KafkaTrackingOptions.cs
@@ -1,0 +1,16 @@
+namespace TestTrackingDiagrams.Extensions.Kafka;
+
+public record KafkaTrackingOptions
+{
+    public string ServiceName { get; set; } = "Kafka";
+    public string CallingServiceName { get; set; } = "Caller";
+    public KafkaTrackingVerbosity Verbosity { get; set; } = KafkaTrackingVerbosity.Detailed;
+    public Func<(string Name, string Id)>? CurrentTestInfoFetcher { get; set; }
+    public Func<string?>? CurrentStepTypeFetcher { get; set; }
+    public bool TrackProduce { get; set; } = true;
+    public bool TrackConsume { get; set; } = true;
+    public bool LogMessageValue { get; set; } = true;
+    public bool LogMessageKey { get; set; } = true;
+    public bool TrackSubscribe { get; set; } = false;
+    public bool TrackCommit { get; set; } = false;
+}