+++ src/TestTrackingDiagrams.Extensions.Kafka/KafkaTrackingVerbosity.cs
@@ -1,0 +1,8 @@
+namespace TestTrackingDiagrams.Extensions.Kafka;
+
+public enum KafkaTrackingVerbosity
+{
+    Raw,
+    Detailed,
+    Summarised
+}